backend/requirements.txt
fastapi==0.109.1
uvicorn==0.23.1
python-multipart==0.0.6
langchain==0.1.20
langchain-google-genai==1.0.10
PyPDF2==3.0.1
python-docx==0.8.11
google-generativeai==0.7.0
